# docs/web/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Admin REST API шины данных (§7 ТЗ). Сессии в Redis +\nAPI-токены (Bearer db_*). SPA по embed.FS отдаётся\nфолбэком на index.html для всех путей не из /api/.",
        "title": "Nexus Web API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/tokens": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Каждый видит только свои токены. Возвращает префикс, scopes, last_used_at — без полной строки токена.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Список API-токенов текущего пользователя (§7.14).",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ListTokensResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Plain-строка токена возвращается ровно один раз, потом доступен только префикс. Scopes — подмножество {logs:read, nodes:read, metrics:read, audit:read}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Создать API-токен (§7.14).",
                "parameters": [
                    {
                        "description": "name + scopes + expires_at",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.createTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "token (plain) + info",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.CreateTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tokens/{id}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Удаление допустимо, только если токен уже revoked. Иначе — 404 «not deletable yet».",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Удалить API-токен.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tokens/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "После revoke токен сразу теряет силу, но сама запись остаётся (для аудита).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Отозвать API-токен.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_web_adapter_in_http.CreateTokenResponse": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/internal_web_adapter_in_http.tokenResponse"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "internal_web_adapter_in_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "password_change_required"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "internal_web_adapter_in_http.ListTokensResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_web_adapter_in_http.tokenResponse"
                    }
                }
            }
        },
...
        "internal_web_adapter_in_http.createTokenRequest": {
            "type": "object",
            "required": [
                "name",
                "scopes"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "scopes": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_web_adapter_in_http.tokenResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_used_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "team_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiTokenAuth": {
            "description": "\"Bearer db_\u003c32-байт-base64\u003e\". См. §7.14.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "nexus_session",
            "in": "cookie"
        }
    }
}